Update Steam - Javascript Scraping with PlayWright.py
Worked on more robust render thresholds, extracting title, thumbnail, and game category tags.

Issue is that the steam HTML has changed, making the process difficult for obtaining the title needed - unsure if this is due to a delay (which DOM should be loaded before scraping), if the class attribute is dynamically generated (which is probably the case), or something else I am not seeing. Will try fixing at future date.